passed test1 TODO: -handle test2 overflow exceptions

+++ sources/Fraction.hpp
@@ -9,7 +9,7 @@ private:
 public:
     Fraction(); // empty constructor will declare 0
     Fraction(int, int);
-    Fraction(double);
+    Fraction(float);
     // getters
     int getNumerator();
     int getDenominator();
